fix(task-descriptions): stabilize Mermaid rendering
Render a fixed Mermaid viewport before SVG generation.

Mount Mermaid content at once in read-only and visual editor paths.

Split task-asset flows and tests, extract the browser route, and align Claude tool expectations.
