--- packages/frontend/src/components/features/task-description-editor/task-description-mermaid-previews.ts
@@ -1,8 +1,0 @@
-import { renderMermaidPreviews } from "@/components/ui/markdown-mermaid-previews";
-import type { MermaidPreviews } from "@/components/ui/markdown-mermaid-state";
-import { getTaskDescriptionVisualMermaidSources } from "./task-description-markdown";
-
-export const renderInitialTaskDescriptionMermaidPreviews = (
-  markdown: string,
-): Promise<MermaidPreviews> =>
-  renderMermaidPreviews(getTaskDescriptionVisualMermaidSources(markdown));
